changed id type from int to long

+++ Licenta.Core/Entities/Role.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Licenta.Core.Entities;
 
-public class Role : IdentityRole<int>
+public class Role : IdentityRole<long>
 {
     public virtual ICollection<UserRole> UserRoles { get; set; }
 }
